♻️ [refactor] #205 : 재사용 버튼 컴포넌트를 스타일 컴포넌트로 변경 및 적용
