+++ client/src/components/NicknameTextField/style.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { OutlinedTextField } from "@styles/TextField";
+import { OutlinedTextField } from "@styles/components";
 
 export const TextFieldWrapper = styled.div`
   display: flex;
