+++ client/src/pages/ChallengePage/style.ts
@@ -1,11 +1,12 @@
 import styled from "styled-components";
-import { DefaultTextField } from "@styles/TextField";
+import { DefaultTextField, DefaultButton } from "@styles/components";
 
 export const Wrapper = styled.div`
   padding: 20px 0;
   gap: 30px;
   display: flex;
   flex-direction: column;
+  align-items: center;
   & > div {
     width: 100%;
     height: fit-content;
@@ -16,6 +17,7 @@ export const Wrapper = styled.div`
     align-items: center;
     @media screen and (max-width: ${({ theme }) => theme.MAX_WIDTH.MOBILE}) {
       padding: 30px;
+      border-radius: 0;
     }
   }
 `;
@@ -55,4 +57,8 @@ export const TextField = styled(DefaultTextField)`
   text-align: end;
   font-size: ${({ theme }) => theme.FONT_SIZE.SMALL};
   font-weight: ${({ theme }) => theme.FONT_WEIGHT.BOLD};
+`;
+
+export const SubmitButton = styled(DefaultButton)`
+  width: 80%;
 `;
