+++ client/src/pages/HomePage/style.ts
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 import challengeSrc from "@public/images/btn_challenge.jpeg";
+import { DefaultButton } from "@styles/components";
 
 export const Wrapper = styled.div`
   gap: 20px;
@@ -10,16 +11,7 @@ export const Wrapper = styled.div`
   align-items: center;
 `;
 
-export const Button = styled.button`
-  padding: 10px 30px;
-  border-radius: 15px;
-  font-size: ${({ theme }) => theme.FONT_SIZE.SMALL};
-  font-weight: ${({ theme }) => theme.FONT_WEIGHT.BOLD};
-  color: ${({ theme }) => theme.COLORS.WHITE};
-  transition: all 0.3s cubic-bezier(0, 0, 0.5, 1);
-`;
-
-export const ChallengeButton = styled(Button)`
+export const ChallengeButton = styled(DefaultButton)`
   width: 95%;
   border-radius: 10px;
   display: flex;
@@ -27,6 +19,7 @@ export const ChallengeButton = styled(Button)`
   justify-content: space-between;
   background: url(${challengeSrc}) no-repeat center;
   background-size: cover;
+  transition: all 0.3s cubic-bezier(0, 0, 0.5, 1);
   filter: saturate(0.3);
   &:hover {
     box-shadow: rgb(0 0 0 / 20%) 2px 4px 12px;
@@ -35,12 +28,12 @@ export const ChallengeButton = styled(Button)`
   }
 `;
 
-export const RecordButton = styled(Button)`
+export const RecordButton = styled(DefaultButton)`
   position: fixed;
   bottom: 80px;
   background-color: ${({ theme }) => theme.COLORS.LIGHT_BLUE};
   box-shadow: rgb(0 0 0 / 30%) 2px 4px 12px;
-
+  transition: all 0.3s cubic-bezier(0, 0, 0.5, 1);
   &:hover {
     background-color: ${({ theme }) => theme.COLORS.HOVER_BLUE};
     box-shadow: rgb(0 0 0 / 35%) 2px 4px 12px;
