--- client/src/styles/TextField.ts
@@ -1,25 +1,0 @@
-import styled from "styled-components";
-
-export const DefaultTextField = styled.input`
-  width: 100%;
-  border: none;
-  caret-color: ${({ theme }) => theme.COLORS.LIGHT_BLUE};
-  &:focus {
-    outline: none;
-  }
-`;
-
-export const OutlinedTextField = styled.input`
-  width: 100%;
-  padding: 15px 20px;
-  border-radius: 10px;
-  border: 1px solid ${({ theme }) => theme.COLORS.LIGHT_GRAY};
-  caret-color: ${({ theme }) => theme.COLORS.LIGHT_BLUE};
-  font-size: ${({ theme }) => theme.FONT_SIZE.LARGE};
-  &:focus {
-    outline-color: ${({ theme }) => theme.COLORS.LIGHT_BLUE};
-  }
-  &::placeholder {
-    color: ${({ theme }) => theme.COLORS.PLACEHOLDER_GRAY};
-  }
-`;
